requirements.txt
pyspark==3.5.1
apache-airflow-providers-apache-spark
